requirements.txt
asgiref==3.4.1
autopep8==1.5.7
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
djongo==1.3.6
dnspython==2.1.0
google-api-core==2.1.1
google-auth==2.3.0
google-cloud-language==2.3.0
googleapis-common-protos==1.53.0
grpcio==1.41.0
gssapi==1.7.2
gunicorn==20.1.0
idna==3.3
mongoengine==0.23.1
packaging==21.0
proto-plus==1.19.5
protobuf==3.19.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==3.0.0
python-snappy==0.6.0
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
twilio==7.2.0
urllib3==1.26.7
whitenoise==5.3.0
